Method-method standard controller

+++ api/database/migrations/2020_12_09_161202_create_items_table.php
@@ -23,7 +23,7 @@ class CreateItemsTable extends Migration
             $table->integer("stock_xl");
             $table->boolean("active")->default(true);
             $table->string("description");
-            $table->double("price");
+            $table->decimal("price", 10, 2);
             $table->text("image64")->nullable();
             $table->timestamps();
         });
